Dynamic font loading implementation & Fixed doUpdate(..)
- BaseObject specification changed: doUpdate(double)
- Font texture is now dynamically loaded.

+++ src/Scene.cpp
@@ -342,7 +342,7 @@ bool Scene::is_input_available() const
   return is_input_available_;
 }
 
-void Scene::doUpdate(float delta)
+void Scene::doUpdate(double delta)
 {
   // scheduled events
   scenetask_.Update(delta);
